# apps/coupon-api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDADAN Coupon API",
    "description": "상점별 쿠폰·도장 발급 시스템 API. 모든 경로는 `/api/coupon/v1` 아래에 있으며, JSON 필드는 snake_case, 식별자는 UUID 문자열, 시각은 RFC 3339 UTC, 금액은 원 단위 정수입니다.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7810",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/coupon/v1/owner/analytics": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "기간별 적립·리워드·캠페인·취소 지표 (§6.3, §19).",
        "description": "Every day in the range appears in the response, including the ones the nightly batch\nhas not reached — those carry `state: PENDING` and no numbers at all, because §19 draws\nthe line between 실시간 수치 and 확정 배치 수치 and a zero would erase it.",
        "operationId": "get_analytics",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "First business day (inclusive)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Last business day (inclusive)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-day and total metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The range is inverted or longer than 366 days"
          },
          "403": {
            "description": "Not a store owner"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AggregationState": {
        "type": "string",
        "description": "Whether a day's numbers are settled (§19: 실시간 수치와 확정 배치 수치를 구분한다).",
        "enum": [
          "PENDING",
          "PROVISIONAL",
          "FINAL"
        ]
      },
      "AnalyticsResponse": {
        "type": "object",
        "description": "The whole answer to `GET /owner/analytics`.",
        "required": [
          "store_id",
          "from",
          "to",
          "finalised_days",
          "pending_days",
          "minimum_cohort_size",
          "totals",
          "days"
        ],
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyMetrics"
            }
          },
          "finalised_days": {
            "type": "integer",
            "format": "int64",
            "description": "Days whose numbers are settled."
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "minimum_cohort_size": {
            "type": "integer",
            "format": "int64",
            "description": "The privacy floor in force (§19)."
          },
          "pending_days": {
            "type": "integer",
            "format": "int64",
            "description": "Days the batch has not reached. Non-zero means the totals below are incomplete, and\nthe client is expected to say so rather than present them as the period's result."
          },
          "store_id": {
            "type": "string",
            "format": "uuid"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totals": {
            "$ref": "#/components/schemas/DailyCounts"
          }
        }
      },
...
      "DailyCounts": {
        "type": "object",
        "description": "§6.3's dashboard row. Gross and reversal are separate columns on purpose.",
        "required": [
          "stamp_earned_count",
          "stamp_voided_count",
          "stamp_net_count",
          "stamp_transaction_count",
          "reward_issued_count",
          "reward_used_count",
          "campaign_coupon_issued_count",
          "campaign_coupon_used_count",
          "campaign_coupon_revoked_count",
          "coupon_expired_count",
          "redemption_voided_count",
          "discount_amount_total"
        ],
        "properties": {
          "active_customer_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Distinct customers seen. Withheld below the privacy floor."
          },
          "campaign_coupon_issued_count": {
            "type": "integer",
            "format": "int64"
          },
          "campaign_coupon_revoked_count": {
            "type": "integer",
            "format": "int64"
          },
          "campaign_coupon_used_count": {
            "type": "integer",
            "format": "int64"
          },
          "coupon_expired_count": {
            "type": "integer",
            "format": "int64"
          },
          "discount_amount_total": {
            "type": "integer",
            "format": "int64"
          },
          "redemption_voided_count": {
            "type": "integer",
            "format": "int64"
          },
          "reward_issued_count": {
            "type": "integer",
            "format": "int64"
          },
          "reward_used_count": {
            "type": "integer",
            "format": "int64"
          },
          "stamp_earned_count": {
            "type": "integer",
            "format": "int64",
            "description": "총 발생치: accruals recorded, before reversals."
          },
          "stamp_net_count": {
            "type": "integer",
            "format": "int64",
            "description": "순수치."
          },
          "stamp_transaction_count": {
            "type": "integer",
            "format": "int64"
          },
          "stamp_voided_count": {
            "type": "integer",
            "format": "int64",
            "description": "취소·보정으로 되돌린 수."
          }
        }
      },
      "DailyMetrics": {
        "type": "object",
        "description": "One business day for one store.",
        "required": [
          "business_day",
          "state",
          "suppressed"
        ],
        "properties": {
          "business_day": {
            "type": "string",
            "format": "date"
          },
          "computed_through": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the figures were last recomputed."
          },
          "metrics": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DailyCounts",
                "description": "`null` while [`AggregationState::Pending`], so a client cannot render a zero."
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/AggregationState"
          },
          "suppressed": {
            "type": "boolean",
            "description": "True when the cohort behind this day is below the privacy floor and the breakdown\nhas been withheld (§19 소규모 집단)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token"
      }
    }
  },
...
}